emulation/python-requirements.txt
coloredlogs
PyYAML
Unidecode
